Handle negative decimals in AVRO parsing.
BigQuery outputs numbers in 128-bit big-endian arrays. Our values never
get quite that large but do end up near 64-bits becase 10^9 digits are
used for decimals since the scale=9 in bigquery decimals.

Thus we manually parse the lower 64-bits of the and compose a javascript
number from it.

Before this, the sample code only worked with positives probably because
there's some weird loop-hole or logic in buffer.readIntBE(off, length)
that skips leading zeros but NOT sign-extensions if length is too big.

This this manually is better. If necessary, we can also parse out
another few bytes in the 128-bit integer. Currently we just check that
they are sign-extension only and throw if not.